slightly improve matching of white pieces
this is still missing two pieces on the test image, but it's better than
before where it matched nothing. it's completely unclear to me why it
wouldn't match with the same values used in python?!

on the `whites*.jpg` test images _all_ red squares are also matched as
white, which is of course wrong!