implement calibration
this now calculates the height difference between men and kings. for
this to work a man must be placed in the top left red field and a king
in the top right red field, equally on the bottom row.

currently the calibration only works with black pieces since the
recognition of white pieces is still not working correctly. however,
this problem is orthogonal and can thus be ignored here.

what's missing is a way to specify the calibration when loading a single
image.