use blue instead of red for dark fields
this solves the problem that the colour of the red fields and of the
white stones (which weren't really white) were too close and thus the
colour segmentation sometimes mistook red fields as white stones.

also add new example images for this. since they have the stones needed
for calibration on the right fields they don't need a separate
calibration file and instead rely on auto-calibration.